\name{mct-package}
\alias{mct-package}
\alias{mct}
\docType{package}
\title{
  Map Comparison Tools
}
\description{
  This is a package with functionalities to quantify the changes between two spatial maps (two raster maps).
}
\details{
  Several methods have been implemented to quantify the changes between two maps.
}
\author{
Babak Naimi

Maintainer: Babak Naimi <naimi.b@gmail.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
